refactor: use default feed
